requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
<<<<<<< HEAD
cffi==1.16.0
=======
>>>>>>> main
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
<<<<<<< HEAD
multidict==6.0.5
orjson==3.10.3
prompt_toolkit==3.0.46
pycparser==2.22
=======
numpy==1.26.4
orjson==3.10.3
pandas==2.2.2
>>>>>>> main
pydantic==2.7.3
pydantic-extra-types==2.8.0
pydantic-settings==2.3.1
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
<<<<<<< HEAD
simple-websocket==1.0.0
=======
>>>>>>> main
six==1.16.0
sniffio==1.3.1
spotipy==2.24.0
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
aioredis
torch
PyJWT
aioredis==2.0.1
numpy